initial commit
The fruits classifying project I did in the summer with my classmates. As the conduct and main coder, we decided using the keras-yolo3(https://github.com/qqwweee/keras-yolo3)in github to make it work. Of course, based on the original projects we wrote our own code to make it. Finally, we successfully build a labeled dataset on our own, and achieved good results.